Added suggestion feature
Any message sent to the bot will now be stored in the database as a feature, and will then have a separate message sent to a provided suggestion channel. In that suggestion channel, the message sent can be reacted to, which will send the reactor a message telling them to respond to the suggestor. Once a message has been received, the bot will then send that information to the suggestor through DMs.